Prior to the below updates, a Heroku deployment was created on the Heroku website and the Parse dashboard was set up on a local terminal session.
Both app:build.gradle and whatsthewait:build.gradle are updated to include Parse dependencies.
AndroidManifest.xml is updated to allow permission to the internet and the device's network state. Also changes the main, first activity to be LoginActivity.java instead of MainActivity.java
Parse is initialized inside the ParseApplication.java file using the configuration variables created during the Heroku deployment.
The RestaurantItem class is annotated with @ParseClassName("RestaurantItem") to allow for persistence.
LoginActivity.java is updated with the logic for logging in a ParseUser and routing them to HomeFragment. An initial check on the currently logged in ParseUser is performed and if a user is already logged in, they are routed immediately to HomeFragment.
SignupActivity.java is updated with the logic for signing up a ParseUser, logging them in, then routing them to HomeFragment. An initial check on the currently logged in ParseUser is performed and if a user is already logged in, they are routed immediately to HomeFragment.
The log out button in ProfileFragment.java is updated with the logic for logging out the current ParseUser. Note: upon completion of the logout functionality, the fragment calls finish on its activity so that a user cannot press the back button to navigate to the previous screen after they have logged out.

Sign up, login, and log out functionality has been tested and works as desired, as shown in this gif (also located in demos/Parse_Login_Signup_Logout_TestPass.gif): https://imgur.com/a/ejzy5FU